requirement.txt
django
djangorestframework
django-taggit
